Refactor get_pitch to audio_framework
